requirements.txt
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
Flask==1.1.1
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
Naked==0.1.31
pycryptodome==3.9.7
PyYAML==5.3.1
requests==2.23.0
shellescape==3.8.1
urllib3==1.25.8
Werkzeug==1.0.0
